pyproject.toml
[tool.poetry]
name = "tabletop-rpg-back"
version = "0.1.0"
description = ""
authors = ["Rémi DUPUY"]
license = "GPL 3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
flask-cors = "^5.0.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.0"
flask-jwt-extended = "^4.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
